Fix Module Name
The module name when building via CocoaPods has been "cdTaskQueue".
When building with Carthage however, it equals "TaskQueue". In order
to make it more easy to import this library for client which itself
will be either managed by Carthage or CocoaPods, the name has been
unified.

The module name for the framework is now "TaskQueue".